requirements.txt
pyautogui
keyboard
interception-python
psutil
time
opencv